Fix unclickable observations bug
